started CI file, added eslint and prettier rules and ran both. eslint needs more config to work properly with typescript
