+++ App.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
-import { useLocalSearchParams } from 'expo-router';
-import { SafeAreaView, StyleSheet, Platform, StatusBar, Text, View } from 'react-native';
-import { Provider } from "react-redux";
+import { SafeAreaView, Text } from 'react-native';
+import { Provider } from 'react-redux';
 import { PaperProvider } from 'react-native-paper';
-import store from "./store";
+import store from './store';
 import rootTheme from './rootTheme';
 
 export default function App() {
@@ -15,6 +14,5 @@ export default function App() {
         </SafeAreaView>
       </Provider>
     </PaperProvider>
-
   );
 }
