+++ app/index.tsx
@@ -1,13 +1,5 @@
 import React from 'react';
-import { useLocalSearchParams } from 'expo-router';
-import {
-  SafeAreaView,
-  StyleSheet,
-  Platform,
-  StatusBar,
-  Text,
-  View,
-} from 'react-native';
+import { SafeAreaView, Text } from 'react-native';
 import { Provider } from 'react-redux';
 import { PaperProvider } from 'react-native-paper';
 import store from '../store';
