+++ store.tsx
@@ -4,13 +4,13 @@ import rootReducer from './reducers/rootReducer';
 import { RootState } from './reducers/rootReducer';
 
 const store = configureStore({
-    reducer: rootReducer
+  reducer: rootReducer,
 });
 
 export type AppDispatch = typeof store.dispatch;
 export const useAppDispatch = () => useDispatch<AppDispatch>();
 
 export type AppSelector = typeof store.getState;
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
 
 export default store;
