+++ tsconfig.json
@@ -2,5 +2,6 @@
   "extends": "expo/tsconfig.base",
   "compilerOptions": {
     "strict": true
-  }
+  },
+  "exclude": ["__tests__/**/*-test.ts"]
 }
